requirements.txt
torch
faiss-gpu
PIL
pandas
ultralytics
opencv-python
tqdm
numpy
streamlit
